+++ esbuild.mjs
@@ -8,6 +8,7 @@ const context = await esbuild.context({
   external: ["obsidian", "electron", "@codemirror/autocomplete", "@codemirror/collab", "@codemirror/commands", "@codemirror/language", "@codemirror/lint", "@codemirror/search", "@codemirror/state", "@codemirror/view", "@lezer/common", "@lezer/highlight", "@lezer/lr"],
   format: "cjs",
   target: "es2022",
+  loader: { ".woff2": "binary" },
   logLevel: "info",
   sourcemap: production ? false : "inline",
   treeShaking: true,